@@ -1,7 +1,7 @@
 [package]
 name = "pueue"
 description = "A cli tool for managing long running shell commands."
-version = "1.0.6"
+version = "2.0.0"
 authors = ["Arne Beer <contact@arne.beer>"]
 homepage = "https://github.com/nukesor/pueue"
 repository = "https://github.com/nukesor/pueue"
@@ -9,18 +9,21 @@ documentation = "https://docs.rs/pueue/"
 license = "MIT"
 keywords = ["shell", "command", "parallel", "task", "queue"]
 readme = "README.md"
-edition = "2018"
+rust-version = "1.58"
+edition = "2021"
 
 [badges]
 maintenance = { status="actively-developed" }
+
+[workspace]
+members = ["lib"]
 
 [[bin]]
 name = "pueue"
 path = "client/main.rs"
 
 # Split the daemon logic in a lib and bin crate.
-# The bin crate uses the lib to for all logic, but using a lib allows
-# us to do proper integration testing.
+# Using a lib allows us to do proper integration testing.
 [lib]
 name = "pueue_daemon_lib"
 path = "daemon/lib.rs"
@@ -30,9 +33,7 @@ name = "pueued"
 path = "daemon/main.rs"
 
 [dependencies]
-pueue-lib = "0.18.0"
-#pueue-lib = { git = "https://github.com/Nukesor/pueue-lib", branch = "master" }
-#pueue-lib = { path = "../../libraries/pueue-lib" }
+pueue-lib = { version="0.19.3", path="lib" }
 
 anyhow = "1"
 chrono = { version="0.4", features=["serde"] }
@@ -41,18 +42,18 @@ shell-escape = "0.1"
 tempfile = "3"
 
 crossbeam-channel = "0.5"
-tokio = { version="1", features=["rt-multi-thread", "time", "macros", "io-std", "io-util"] }
+tokio = { version="1", features=["rt-multi-thread", "time", "io-std"] }
 snap = "1"
 serde = "1"
 serde_json = "1"
 serde_derive = "1"
 
 log = "0.4"
-simplelog = { version="0.10", default-features=false }
+simplelog = { version="0.11", default-features=false }
 clap = { version="3", features=["derive", "cargo"] }
 clap_complete = "3"
-crossterm = "0.20"
-comfy-table = "4"
+crossterm = "0.23"
+comfy-table = "5"
 
 handlebars = "4"
 ctrlc = { version="3", features=["termination"] }
@@ -69,19 +70,16 @@ winapi = { version="0.3", features=[
 
 # Unix
 [target.'cfg(not(windows))'.dependencies]
-nix = "0.22"
+nix = "0.23"
 whoami = "1"
 
 # Linux / Freebsd
 [target.'cfg(any(target_os = "linux", target_os = "freebsd"))'.dependencies]
-procfs = { version="0.9", default-features=false }
-
-[build-dependencies]
-version_check = "0.9"
+procfs = { version="0.12", default-features=false }
 
 [dev-dependencies]
 assert_cmd = "2"
 anyhow = "1"
-better-panic = "0.2"
-pretty_assertions = "0.7"
-rstest = "0.11"
+better-panic = "0.3"
+pretty_assertions = "1"
+rstest = "0.12"
