@@ -4,10 +4,17 @@ use pueue_lib::network::message::Message;
 use pueue_lib::network::protocol::*;
 use pueue_lib::state::State;
 
-pub mod edit;
-pub mod local_follow;
-pub mod restart;
-pub mod wait;
+mod edit;
+mod format_state;
+mod local_follow;
+mod restart;
+mod wait;
+
+pub use edit::edit;
+pub use format_state::format_state;
+pub use local_follow::local_follow;
+pub use restart::restart;
+pub use wait::wait;
 
 // This is a helper function for easy retrieval of the current daemon state.
 // The current daemon state is often needed in more complex commands.
