@@ -1,9 +1,14 @@
 use std::path::PathBuf;
 
-use clap::Parser;
+use clap::{Parser, ValueHint};
 
 #[derive(Parser, Debug)]
-#[clap(name = "Pueue daemon", about, author, version)]
+#[clap(
+    name = "Pueue daemon",
+    about = "Start the Pueue daemon",
+    author,
+    version
+)]
 pub struct CliArguments {
     /// Verbose mode (-v, -vv, -vvv)
     #[clap(short, long, parse(from_occurrences))]
@@ -17,6 +22,10 @@ pub struct CliArguments {
 
     /// Path to a specific pueue config file to use.
     /// This ignores all other config files.
-    #[clap(short, long)]
+    #[clap(short, long, value_hint = ValueHint::FilePath)]
     pub config: Option<PathBuf>,
+
+    /// The name of the profile that should be loaded from your config file.
+    #[clap(short, long)]
+    pub profile: Option<String>,
 }
