@@ -9,10 +9,11 @@ use crate::network::response_helper::{ensure_group_exists, task_action_response_
 /// Invoked when calling `pueue kill`.
 /// Forward the kill message to the task handler, which then kills the process.
 pub fn kill(message: KillMessage, sender: &Sender<Message>, state: &SharedState) -> Message {
-    let state = state.lock().unwrap();
+    let mut state = state.lock().unwrap();
+
     // If a group is selected, make sure it exists.
     if let TaskSelection::Group(group) = &message.tasks {
-        if let Err(message) = ensure_group_exists(&state, group) {
+        if let Err(message) = ensure_group_exists(&mut state, group) {
             return message;
         }
     }
@@ -30,11 +31,10 @@ pub fn kill(message: KillMessage, sender: &Sender<Message>, state: &SharedState)
                 &state,
             ),
             TaskSelection::Group(group) => create_success_message(format!(
-                "Sending signal {} to all running tasks of group {}.",
-                signal, group
+                "Sending signal {signal} to all running tasks of group {group}.",
             )),
             TaskSelection::All => {
-                create_success_message(format!("Sending signal {} to all running tasks.", signal))
+                create_success_message(format!("Sending signal {signal} to all running tasks."))
             }
         }
     } else {
@@ -46,10 +46,11 @@ pub fn kill(message: KillMessage, sender: &Sender<Message>, state: &SharedState)
                 &state,
             ),
             TaskSelection::Group(group) => create_success_message(format!(
-                "All tasks of group \"{}\" are being killed.",
-                group
+                "All tasks of group \"{group}\" are being killed. The group will also be paused!!!"
             )),
-            TaskSelection::All => create_success_message("All tasks are being killed."),
+            TaskSelection::All => {
+                create_success_message("All tasks are being killed. All groups will be paused!!!")
+            }
         }
     }
 }
