@@ -1,4 +1,4 @@
-use std::collections::BTreeMap;
+use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use std::process::{Child, Command, Stdio};
 
@@ -10,22 +10,24 @@ use tokio::io::{self, AsyncWriteExt};
 use pueue_daemon_lib::run;
 use pueue_lib::settings::*;
 
-use super::{get_pid, sleep_ms};
+use crate::helper::*;
+
+/// All info about a booted standalone test daemon.
+/// This daemon is executed in the same async environement as the rest of the test.
+pub struct PueueDaemon {
+    pub settings: Settings,
+    pub tempdir: TempDir,
+    pub pid: i32,
+}
 
 /// A helper function which, creates some test config, sets up a temporary directory and boots a
 /// daemon in a async tokio thread.
 /// This is done in 90% of our tests, thereby this convenience helper.
-pub fn threaded_setup() -> Result<(Settings, TempDir, i32)> {
-    let (settings, tempdir) = base_setup()?;
-    let pid = boot_daemon(tempdir.path())?;
+pub async fn daemon() -> Result<PueueDaemon> {
+    let (settings, tempdir) = daemon_base_setup()?;
 
-    Ok((settings, tempdir, pid))
-}
-
-/// Spawn the daemon main logic in it's own async function.
-/// It'll be executed by the tokio multi-threaded executor.
-pub fn boot_daemon(pueue_dir: &Path) -> Result<i32> {
-    let path = pueue_dir.clone().to_path_buf();
+    let pueue_dir = tempdir.path();
+    let path = pueue_dir.to_path_buf();
     // Start/spin off the daemon and get its PID
     tokio::spawn(run_and_handle_error(path, true));
     let pid = get_pid(pueue_dir)?;
@@ -38,7 +40,12 @@ pub fn boot_daemon(pueue_dir: &Path) -> Result<i32> {
     while current_try < tries {
         sleep_ms(50);
         if socket_path.exists() {
-            return Ok(pid);
+            create_test_groups(&settings.shared).await?;
+            return Ok(PueueDaemon {
+                settings,
+                tempdir,
+                pid,
+            });
         }
 
         current_try += 1;
@@ -47,14 +54,31 @@ pub fn boot_daemon(pueue_dir: &Path) -> Result<i32> {
     bail!("Daemon didn't boot after 1sec")
 }
 
+/// Internal helper function, which wraps the daemon main logic inside tokio and prints any errors.
+async fn run_and_handle_error(pueue_dir: PathBuf, test: bool) -> Result<()> {
+    if let Err(err) = run(Some(pueue_dir.join("pueue.yml")), None, test).await {
+        let mut stdout = io::stdout();
+        stdout
+            .write_all(format!("Entcountered error: {:?}", err).as_bytes())
+            .await
+            .expect("Failed to write to stdout.");
+        stdout.flush().await?;
+
+        return Err(err);
+    }
+
+    Ok(())
+}
+
 /// Spawn the daemon by calling the actual pueued binary.
-/// This function also checks for the pid file and the unix socket to pop-up.
-pub fn boot_standalone_daemon(pueue_dir: &Path) -> Result<Child> {
+/// This function also checks for the pid file and the unix socket to appear.
+pub async fn standalone_daemon(pueue_dir: &Path) -> Result<Child> {
     let child = Command::cargo_bin("pueued")?
         .arg("--config")
         .arg(pueue_dir.join("pueue.yml").to_str().unwrap())
         .arg("-vvv")
         .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
         .spawn()?;
 
     let tries = 20;
@@ -74,23 +98,8 @@ pub fn boot_standalone_daemon(pueue_dir: &Path) -> Result<Child> {
     bail!("Daemon didn't boot in stand-alone mode after 1sec")
 }
 
-/// Internal helper function, which wraps the daemon main logic and prints any errors.
-async fn run_and_handle_error(pueue_dir: PathBuf, test: bool) -> Result<()> {
-    if let Err(err) = run(Some(pueue_dir.join("pueue.yml")), test).await {
-        let mut stdout = io::stdout();
-        stdout
-            .write_all(format!("Entcountered error: {:?}", err).as_bytes())
-            .await
-            .expect("Failed to write to stdout.");
-        stdout.flush().await?;
-
-        return Err(err);
-    }
-
-    Ok(())
-}
-
-pub fn base_setup() -> Result<(Settings, TempDir)> {
+/// This is the base setup for all daemon test setups.
+pub fn daemon_base_setup() -> Result<(Settings, TempDir)> {
     // Create a temporary directory used for testing.
     let tempdir = TempDir::new().unwrap();
     let tempdir_path = tempdir.path();
@@ -98,16 +107,17 @@ pub fn base_setup() -> Result<(Settings, TempDir)> {
     std::fs::create_dir(tempdir_path.join("certs")).unwrap();
 
     let shared = Shared {
-        pueue_directory: tempdir_path.clone().to_path_buf(),
+        pueue_directory: Some(tempdir_path.to_path_buf()),
+        runtime_directory: Some(tempdir_path.to_path_buf()),
         #[cfg(not(target_os = "windows"))]
         use_unix_socket: true,
         #[cfg(not(target_os = "windows"))]
-        unix_socket_path: tempdir_path.join("test.socket"),
+        unix_socket_path: Some(tempdir_path.join("test.socket")),
         host: "localhost".to_string(),
         port: "51230".to_string(),
-        daemon_cert: tempdir_path.join("certs").join("daemon.cert"),
-        daemon_key: tempdir_path.join("certs").join("daemon.key"),
-        shared_secret_path: tempdir_path.join("secret"),
+        daemon_cert: Some(tempdir_path.join("certs").join("daemon.cert")),
+        daemon_key: Some(tempdir_path.join("certs").join("daemon.key")),
+        shared_secret_path: Some(tempdir_path.join("secret")),
     };
 
     let client = Client {
@@ -121,25 +131,20 @@ pub fn base_setup() -> Result<(Settings, TempDir)> {
         status_datetime_format: "%Y-%m-%d\n%H:%M:%S".into(),
     };
 
-    let mut groups = BTreeMap::new();
-    groups.insert(PUEUE_DEFAULT_GROUP.to_string(), 1);
-    groups.insert("test_2".to_string(), 2);
-    groups.insert("test_3".to_string(), 3);
-    groups.insert("test_5".to_string(), 5);
-
+    #[allow(deprecated)]
     let daemon = Daemon {
-        default_parallel_tasks: 1,
         pause_group_on_failure: false,
         pause_all_on_failure: false,
         callback: None,
         callback_log_lines: 15,
-        groups,
+        groups: None,
     };
 
     let settings = Settings {
         client,
         daemon,
         shared,
+        profiles: HashMap::new(),
     };
 
     settings
@@ -147,4 +152,16 @@ pub fn base_setup() -> Result<(Settings, TempDir)> {
         .context("Couldn't write pueue config to temporary directory")?;
 
     Ok((settings, tempdir))
+}
+
+/// Create a few test groups that have various parallel task settings.
+pub async fn create_test_groups(shared: &Shared) -> Result<()> {
+    add_group_with_slots(shared, "test_2", 2).await?;
+    add_group_with_slots(shared, "test_3", 3).await?;
+    add_group_with_slots(shared, "test_5", 5).await?;
+
+    wait_for_group(shared, "test_3").await?;
+    wait_for_group(shared, "test_5").await?;
+
+    Ok(())
 }
