@@ -2,20 +2,23 @@ use anyhow::Result;
 
 use pueue_lib::network::message::*;
 
+use crate::fixtures::*;
 use crate::helper::*;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 /// Add and directly remove a group.
 async fn test_add_and_remove() -> Result<()> {
-    let (settings, tempdir) = base_setup()?;
-    let shared = &settings.shared;
-    let _pid = boot_daemon(tempdir.path())?;
+    let daemon = daemon().await?;
+    let shared = &daemon.settings.shared;
 
     // Add a new group
     add_group_with_slots(shared, "testgroup", 1).await?;
 
     // Try to add the same group again. This should fail
-    let add_message = Message::Group(GroupMessage::Add("testgroup".to_string()));
+    let add_message = Message::Group(GroupMessage::Add {
+        name: "testgroup".to_string(),
+        parallel_tasks: None,
+    });
     assert_failure(send_message(shared, add_message).await?);
 
     // Remove the newly added group and wait for the deletion to be processed.
@@ -33,12 +36,10 @@ async fn test_add_and_remove() -> Result<()> {
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 /// Users cannot delete the default group.
 async fn test_cannot_delete_default() -> Result<()> {
-    let (settings, tempdir) = base_setup()?;
-    let shared = &settings.shared;
-    let _pid = boot_daemon(tempdir.path())?;
+    let daemon = daemon().await?;
 
     let pause_message = Message::Group(GroupMessage::Remove(PUEUE_DEFAULT_GROUP.to_string()));
-    assert_failure(send_message(shared, pause_message).await?);
+    assert_failure(send_message(&daemon.settings.shared, pause_message).await?);
 
     Ok(())
 }
@@ -46,12 +47,10 @@ async fn test_cannot_delete_default() -> Result<()> {
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 /// Users cannot delete a non-existing group.
 async fn test_cannot_delete_non_existing() -> Result<()> {
-    let (settings, tempdir) = base_setup()?;
-    let shared = &settings.shared;
-    let _pid = boot_daemon(tempdir.path())?;
+    let daemon = daemon().await?;
 
     let pause_message = Message::Group(GroupMessage::Remove("doesnt_exist".to_string()));
-    assert_failure(send_message(shared, pause_message).await?);
+    assert_failure(send_message(&daemon.settings.shared, pause_message).await?);
 
     Ok(())
 }
@@ -59,16 +58,15 @@ async fn test_cannot_delete_non_existing() -> Result<()> {
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 /// Groups with tasks shouldn't be able to be removed.
 async fn test_cannot_delete_group_with_tasks() -> Result<()> {
-    let (settings, tempdir) = base_setup()?;
-    let shared = &settings.shared;
-    let _pid = boot_daemon(tempdir.path())?;
+    let daemon = daemon().await?;
+    let shared = &daemon.settings.shared;
 
     // Add a new group
     add_group_with_slots(shared, "testgroup", 1).await?;
 
     // Add a task
-    assert_success(fixtures::add_task_to_group(shared, "ls", "testgroup").await?);
-    wait_for_task_condition(&settings.shared, 0, |task| task.is_done()).await?;
+    assert_success(add_task_to_group(shared, "ls", "testgroup").await?);
+    wait_for_task_condition(&daemon.settings.shared, 0, |task| task.is_done()).await?;
 
     // We shouldn't be capable of removing that group
     let pause_message = Message::Group(GroupMessage::Remove("testgroup".to_string()));
