@@ -3,14 +3,14 @@ use pueue_lib::network::message::*;
 use pueue_lib::state::GroupStatus;
 use pueue_lib::task::*;
 
-use crate::helper::fixtures::add_task;
+use crate::fixtures::*;
 use crate::helper::*;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 /// Make sure that no tasks will be started in a paused queue
 async fn test_pause_daemon() -> Result<()> {
-    let (settings, _tempdir, _pid) = threaded_setup()?;
-    let shared = &settings.shared;
+    let daemon = daemon().await?;
+    let shared = &daemon.settings.shared;
 
     // This pauses the daemon
     pause_tasks(shared, TaskSelection::All).await?;
@@ -30,8 +30,8 @@ async fn test_pause_daemon() -> Result<()> {
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 /// Make sure that running tasks will be properly paused
 async fn test_pause_running_task() -> Result<()> {
-    let (settings, _tempdir, _pid) = threaded_setup()?;
-    let shared = &settings.shared;
+    let daemon = daemon().await?;
+    let shared = &daemon.settings.shared;
 
     // Start a long running task and make sure it's started
     add_task(shared, "sleep 60", false).await?;
@@ -44,8 +44,8 @@ async fn test_pause_running_task() -> Result<()> {
     wait_for_task_condition(shared, 0, |task| matches!(task.status, TaskStatus::Paused)).await?;
     let state = get_state(shared).await?;
     assert_eq!(
-        state.groups.get(PUEUE_DEFAULT_GROUP).unwrap(),
-        &GroupStatus::Paused
+        state.groups.get(PUEUE_DEFAULT_GROUP).unwrap().status,
+        GroupStatus::Paused
     );
 
     Ok(())
@@ -54,8 +54,8 @@ async fn test_pause_running_task() -> Result<()> {
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 /// A queue can get paused, while the tasks may finish on their own.
 async fn test_pause_with_wait() -> Result<()> {
-    let (settings, _tempdir, _pid) = threaded_setup()?;
-    let shared = &settings.shared;
+    let daemon = daemon().await?;
+    let shared = &daemon.settings.shared;
 
     // Start a long running task and make sure it's started
     add_task(shared, "sleep 60", false).await?;
