@@ -1,13 +1,14 @@
 use anyhow::Result;
 use pueue_lib::network::message::*;
 
+use crate::fixtures::*;
 use crate::helper::*;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-/// Ensure that clean only removes finished tasks
-async fn test_normal_clean() -> Result<()> {
-    let (settings, _tempdir, _pid) = threaded_setup()?;
-    let shared = &settings.shared;
+/// Ensure that only removable tasks can be removed.
+async fn test_normal_remove() -> Result<()> {
+    let daemon = daemon().await?;
+    let shared = &daemon.settings.shared;
 
     // We'll add some tasks.
     // Task 0-2 will be immediately handled by the daemon, the other three tasks are queued for
@@ -18,8 +19,8 @@ async fn test_normal_clean() -> Result<()> {
     // 3 -> paused
     // 4 -> queued
     // 5 -> stashed
-    for command in vec!["failing", "ls", "sleep 60", "sleep 60", "ls", "ls"] {
-        assert_success(fixtures::add_task(shared, command, false).await?);
+    for command in &["failing", "ls", "sleep 60", "sleep 60", "ls", "ls"] {
+        assert_success(add_task(shared, command, false).await?);
     }
     // Wait for task2 to start. This implies task[0,1] being finished.
     wait_for_task_condition(shared, 2, |task| task.is_running()).await?;
@@ -28,7 +29,7 @@ async fn test_normal_clean() -> Result<()> {
     start_tasks(shared, TaskSelection::TaskIds(vec![3])).await?;
     wait_for_task_condition(shared, 3, |task| task.is_running()).await?;
 
-    pause_tasks(&shared, TaskSelection::TaskIds(vec![3])).await?;
+    pause_tasks(shared, TaskSelection::TaskIds(vec![3])).await?;
 
     // Stash task 5
     let pause_message = Message::Stash(vec![5]);
