@@ -7,28 +7,29 @@ use pretty_assertions::assert_eq;
 use pueue_lib::network::message::TaskSelection;
 use pueue_lib::state::GroupStatus;
 
+use crate::fixtures::*;
 use crate::helper::*;
 
 #[tokio::test]
 /// The daemon should start in the same state as before shutdown, if no tasks are queued.
 /// This function tests for the running state.
 async fn test_start_running() -> Result<()> {
-    let (settings, tempdir) = base_setup()?;
-    let child = boot_standalone_daemon(tempdir.path())?;
+    let (settings, tempdir) = daemon_base_setup()?;
+    let child = standalone_daemon(tempdir.path()).await?;
     let shared = &settings.shared;
 
     // Kill the daemon and wait for it to shut down.
-    assert_success(shutdown_daemon(&shared).await?);
+    assert_success(shutdown_daemon(shared).await?);
     wait_for_shutdown(child.id().try_into()?)?;
 
     // Boot it up again
-    let mut child = boot_standalone_daemon(tempdir.path())?;
+    let mut child = standalone_daemon(tempdir.path()).await?;
 
     // Assert that the group is still running.
     let state = get_state(shared).await?;
     assert_eq!(
-        state.groups.get(PUEUE_DEFAULT_GROUP).unwrap(),
-        &GroupStatus::Running
+        state.groups.get(PUEUE_DEFAULT_GROUP).unwrap().status,
+        GroupStatus::Running
     );
 
     child.kill()?;
@@ -39,25 +40,25 @@ async fn test_start_running() -> Result<()> {
 /// The daemon should start in the same state as before shutdown, if no tasks are queued.
 /// This function tests for the paused state.
 async fn test_start_paused() -> Result<()> {
-    let (settings, tempdir) = base_setup()?;
-    let child = boot_standalone_daemon(tempdir.path())?;
+    let (settings, tempdir) = daemon_base_setup()?;
+    let child = standalone_daemon(tempdir.path()).await?;
     let shared = &settings.shared;
 
     // This pauses the daemon
     pause_tasks(shared, TaskSelection::All).await?;
 
     // Kill the daemon and wait for it to shut down.
-    assert_success(shutdown_daemon(&shared).await?);
+    assert_success(shutdown_daemon(shared).await?);
     wait_for_shutdown(child.id().try_into()?)?;
 
     // Boot it up again
-    let mut child = boot_standalone_daemon(tempdir.path())?;
+    let mut child = standalone_daemon(tempdir.path()).await?;
 
     // Assert that the group is still paused.
     let state = get_state(shared).await?;
     assert_eq!(
-        state.groups.get(PUEUE_DEFAULT_GROUP).unwrap(),
-        &GroupStatus::Paused
+        state.groups.get(PUEUE_DEFAULT_GROUP).unwrap().status,
+        GroupStatus::Paused
     );
 
     child.kill()?;
@@ -67,8 +68,8 @@ async fn test_start_paused() -> Result<()> {
 #[tokio::test]
 /// The daemon will load new settings, when restoring a previous state.
 async fn test_load_config() -> Result<()> {
-    let (mut settings, tempdir) = base_setup()?;
-    let child = boot_standalone_daemon(tempdir.path())?;
+    let (mut settings, tempdir) = daemon_base_setup()?;
+    let child = standalone_daemon(tempdir.path()).await?;
 
     // Kill the daemon and wait for it to shut down.
     assert_success(shutdown_daemon(&settings.shared).await?);
@@ -77,11 +78,10 @@ async fn test_load_config() -> Result<()> {
     // Change the settings and save it to disk
     settings.client.dark_mode = true;
     settings.daemon.callback = Some("This is a test".to_string());
-    settings.shared.daemon_key = PathBuf::from("/tmp/daemon.key");
+    settings.shared.daemon_key = Some(PathBuf::from("/tmp/daemon.key"));
     settings.save(&Some(tempdir.path().join("pueue.yml")))?;
 
-    // Boot it up again
-    let mut child = boot_standalone_daemon(tempdir.path())?;
+    let mut child = standalone_daemon(tempdir.path()).await?;
 
     // Get the new state and make sure the settings actually changed.
     let state = get_state(&settings.shared).await?;
